reduce directory level
